opapi: exclude Online field from Printer struct.
This is a big hack: since we can't really distinguish between
Online == false and Online is not set, our updates from database
clear Online status. To workaround this, we remove Online field from
the struct, and use a separate struct (FakePrinterForOnlineStatus, see the cloud repo),
to set the status.

This is horrible.